Updates * Couple small update to fix port mapping when runnign a docker package
