+++ src/utils/docker/buildDockerCmd.js
@@ -87,7 +87,7 @@ const createRunCmd = (globalConfig, dockerCmd, params) => {
 
   // Temp helper to map port 80 to the DOC_APP_PORT
   // This should be removed when keg-proxy is setup
-  dockerCmd = getPortMap(dockerCmd, context)
+  // dockerCmd = getPortMap(dockerCmd, context)
 
   // Add the env to the docker command
   dockerCmd = addContainerEnv(dockerCmd, {
