+++ src/utils/docker/compose/index.js
@@ -1,10 +1,10 @@
 module.exports = {
+  ...require('./addExposedPorts'),
   ...require('./buildComposeCmd'),
   ...require('./buildServiceName'),
   ...require('./checkKillRunning'),
   ...require('./loadComposeConfig'),
   ...require('./removeInjected'),
   ...require('./getServiceName'),
-  ...require('./getServicePorts'),
   ...require('./getServiceVolumes'),
 }
